Task/fixbug update portfolio (#52)
* update portfolio

* fix bug, upload img

* update img

* small fix

* fix bug

* fix contact, portfolio bug

* fix img 404

* list team back

* back after contact

* fix team detail

* fix build

* fix img at team detail

* update list team page

+++ src/pages/team/[teamId]/dashboard/portfolio/[portfolioId]/index.tsx
@@ -1,16 +1,18 @@
 import { PortfolioApi } from "@/api/portfolio-api";
 import { ManagePortfolio } from "@/src/layouts/manage-team/Manage-portfolio";
 import { EPrivacy, IPortfolio } from "@/type/team/team.type";
-import { DeleteOutline, LockOutlined } from "@mui/icons-material";
+import { ApiOutlined, CalendarMonthOutlined, DeleteOutline, InsertLinkOutlined, LabelOutlined, LockOutlined } from "@mui/icons-material";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 import { toast } from "react-toastify";
+import Image from "next/image"
 
 const PortfolioDetail = () => {
   const [teamId, setTeamId] = useState<number>(NaN);
   const [portfolioId, setPortfolioId] = useState<number>(NaN);
   const [portfolio, setPortfolio] = useState<IPortfolio>({} as IPortfolio);
+  const [ isDeleting, setIsDeleting ] = useState(false);
   const router = useRouter();
   useEffect(() => {
     if (router.query.teamId) {
@@ -37,21 +39,28 @@ const PortfolioDetail = () => {
     }
   },[portfolioId]);
 
+  useEffect(() => {
+    if(isDeleting){
+      PortfolioApi.deletePortfolio(portfolioId)
+        .then((res) => {
+          if (res) {
+            toast.success("Prtfolio delete  successfull");
+            router.push(`/team/${teamId}/dashboard/portfolio`);
+          } else {
+            toast.error("Failed delete portfolio");
+            setIsDeleting(false);
+          }
+        })
+        .catch((err) => {
+          toast.error("Failed delete portfolio");
+          setIsDeleting(false);
+        });
+    }
+  },[isDeleting])
   const handleDelte = () => {
-    PortfolioApi.deletePortfolio(portfolioId)
-      .then((res) => {
-        if (res) {
-          toast.success("success delete portfolio");
-          router.push(`/team/${teamId}/dashboard/portfolio`);
-        } else {
-          toast.error("failed delete portfolio");
-        }
-      })
-      .catch((err) => {
-        toast.error("failed delete portfolio");
-      });
+    setIsDeleting(true);
   };
-  return (
+  return (  
     <ManagePortfolio>
       <div className="">
         <div className="flex items-center justify-between border-b ">
@@ -68,17 +77,26 @@ const PortfolioDetail = () => {
                 target="_blank"
                 rel="noopener noreferrer"
               >
+                <InsertLinkOutlined className="text-sm mr-2" />{" "}
                 {portfolio.clientWebsite}
               </a>
             ) : null}
             {portfolio.category ? (
-              <span className="text-sm">{portfolio.category}</span>
+              <span className="text-sm">
+                {" "}
+                <ApiOutlined className="text-sm mr-2" />
+                {portfolio.category}
+              </span>
             ) : null}
             {portfolio.estimate ? (
-              <span className="text-sm">{portfolio.estimate}</span>
+              <span className="text-sm">
+                <LabelOutlined className="text-sm mr-2" />
+                {portfolio.estimate}
+              </span>
             ) : null}
             {portfolio.startDate ? (
               <span className="text-sm">
+                <CalendarMonthOutlined className="text-sm mr-2" />
                 {portfolio.startDate}
                 {portfolio.endDate ? " to " + portfolio.endDate : null}
               </span>
@@ -88,6 +106,24 @@ const PortfolioDetail = () => {
             <p>{portfolio.description}</p>
           </div>
         </div>
+        {portfolio.imageUrl ? (
+          <div className="w-full flex items-center justify-center">
+            <Image
+              src={
+                "https://kryptohub-be.herokuapp.com/api/portfolio/getImage/" +
+                  portfolio.imageUrl || "/user1.png"
+              }
+              alt="img"
+              width={300}
+              height={300}
+            />
+          </div>
+        ) : null}
+        {portfolio.videoLink ? (
+          <div className="w-full flex items-center justify-center">
+            <a href={portfolio.videoLink}>{portfolio.videoLink}</a>
+          </div>
+        ) : null}
         <div className="flex items-center justify-between p-2">
           <Link href={`#`}>
             <a onClick={handleDelte}>
